feat(authz): itnroduce container for DI, move srvicer to entrypoints/server/servicer.py
