+++ services/authz/entrypoints/server/main.py
@@ -17,80 +17,15 @@ if str(REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(REPO_ROOT))
 
 from shared.pkg.py.authz.v1 import authz_pb2, authz_pb2_grpc
-from internal.rebac.model import PermissionService
-from internal.neo4j.store import Neo4jStore
-from internal.types import Tuple
-from internal.cache.redis_cache import RedisDecisionCache
-from internal.kafka.producer import AuditProducer
-from internal.config import cfg
-
 from shared.pkg.py_kit import logger, metric
 from shared.pkg.py_kit.tracing import init_tracer, shutdown_tracer
 from shared.pkg.py_kit.tracing.grpc_interceptor import TracingServerInterceptor
 from shared.pkg.py_kit.middleware import MetricsServerInterceptor
 from shared.pkg.py_kit.closer import configure_default, add_named
 from internal import metric as authz_metrics
-
-
-
-class PermissionServiceServicer(authz_pb2_grpc.PermissionServiceServicer):
-    """gRPC service implementation"""
-
-    def __init__(self):
-        self._neo4j_store = Neo4jStore(uri=cfg.neo4j_uri(), user=cfg.neo4j_user(), password=cfg.neo4j_password())
-        self._cache = RedisDecisionCache(host=cfg.redis_host(), port=cfg.redis_port())
-        audit_producer = AuditProducer(cfg.kafka_bootstrap())
-        self.rebac = PermissionService(
-            store=self._neo4j_store, cache=self._cache, audit_producer=audit_producer
-        )
-
-    def Check(self, request, context):
-        logger.debug({}, "Check RPC", subject=request.subject, action=request.action, object=request.object)
-        allowed = self.rebac.check(request.subject, request.action, request.object)
-        logger.info({}, "Check RPC done", subject=request.subject, action=request.action, object=request.object, allowed=allowed)
-        return authz_pb2.CheckResponse(
-            allowed=allowed, reason="Neo4j ReBAC (transitive + HAS_PERMISSION)"
-        )
-
-    def WriteTuple(self, request, context):
-        logger.debug({}, "WriteTuple RPC", subject=request.subject, relation=request.relation, object=request.object)
-        level = request.level if request.level else None
-        tuple_ = Tuple(request.subject, request.relation, request.object, level=level)
-        success = self.rebac.write_tuple(tuple_)
-        logger.info({}, "WriteTuple RPC done", success=success)
-        return authz_pb2.WriteTupleResponse(success=success)
-
-    def DeleteTuple(self, request, context):
-        logger.debug({}, "DeleteTuple RPC", subject=request.subject, relation=request.relation, object=request.object)
-        tuple_ = Tuple(request.subject, request.relation, request.object)
-        success = self.rebac.delete_tuple(tuple_)
-        logger.info({}, "DeleteTuple RPC done", success=success)
-        return authz_pb2.DeleteTupleResponse(success=success)
-
-    def Read(self, request, context):
-        logger.debug({}, "Read RPC", subject=request.subject)
-        tuples = self.rebac.read_tuples(request.subject)
-        response = authz_pb2.ReadResponse()
-        for t in tuples:
-            rt = response.tuples.add(subject=t.subject, relation=t.relation, object=t.object)
-            if t.level:
-                rt.level = t.level
-        logger.debug({}, "Read RPC done", count=len(tuples))
-        return response
-
-    def HealthCheck(self, request, context):
-        status = authz_pb2.HealthCheckResponse.SERVING
-        try:
-            self._neo4j_store.driver.verify_connectivity()
-        except Exception as e:
-            logger.warn({}, "Neo4j health check failed", error=str(e))
-            status = authz_pb2.HealthCheckResponse.NOT_SERVING
-        try:
-            self._cache._client.ping()
-        except Exception as e:
-            logger.warn({}, "Redis health check failed", error=str(e))
-            status = authz_pb2.HealthCheckResponse.NOT_SERVING
-        return authz_pb2.HealthCheckResponse(status=status)
+from internal.config import cfg
+from internal.container import Container
+from entrypoints.server.servicer import PermissionServiceServicer
 
 
 def serve():
@@ -108,8 +43,13 @@ def serve():
 
     closer = configure_default(signal.SIGINT, signal.SIGTERM)
 
-    add_named("otel-tracing",  shutdown_tracer)
-    add_named("otel-metrics",  lambda: metric.shutdown())
+    add_named("otel-tracing", shutdown_tracer)
+    add_named("otel-metrics", lambda: metric.shutdown())
+
+    # ── Dependencies ──────────────────────────────────────────────────────
+
+    container = Container(cfg)
+    add_named("neo4j", container.neo4j_store.close)
 
     # ── gRPC server ────────────────────────────────────────────────────────
 
@@ -124,7 +64,7 @@ def serve():
     add_named("grpc-server", lambda: server.stop(grace=5).wait())
 
     authz_pb2_grpc.add_PermissionServiceServicer_to_server(
-        PermissionServiceServicer(), server
+        PermissionServiceServicer(container), server
     )
 
     # gRPC Health
